requirements.txt
# ── 核心資料平台 ──────────────────────────────────────────────
pandas>=2.3.0
numpy>=2.4.0
pytz>=2025.1

# ── 金融資料來源 ──────────────────────────────────────────────
# FinLab：台股回測引擎 + 資料
#   免費版每日 500 MB 額度（https://finlab.tw）
#   VIP 訂閱版每日 5000 MB
finlab>=1.5.7

yfinance>=1.2.0              # 美股行情（免費）

# FinMind（台股資料，Apache 2.0，免費）
# 官方 PyPI 版：pip install finmind
# 或從原始碼安裝：pip install git+https://github.com/FinMind/FinMind.git
# finmind>=0.0.1

# ── Telegram Bot ──────────────────────────────────────────────
python-telegram-bot>=22.6

# ── 排程器 ────────────────────────────────────────────────────
APScheduler>=3.11.0

# ── Claude AI ─────────────────────────────────────────────────
anthropic>=0.83.0

# ── HTTP / 非同步 ─────────────────────────────────────────────
aiohttp>=3.13.0
httpx>=0.28.0
requests>=2.32.0

# ── 工具 ──────────────────────────────────────────────────────
python-dotenv>=1.2.0
loguru>=0.7.0

# ── 量化分析（實驗性策略用）──────────────────────────────────
statsmodels>=0.14.0          # exp_pair_trading_semi.py（配對交易）
ta>=0.5.25                   # 技術指標輔助
